rtc: gamecube: bias is signed for set_time too

+++ drivers/rtc/rtc-gamecube.c
@@ -133,7 +133,7 @@ static int gamecube_rtc_set_time(struct device *dev, struct rtc_time *t)
 
 	/* Subtract the timestamp and the bias to obtain the counter value */
 	timestamp = rtc_tm_to_time64(t);
-	return regmap_write(d->regmap, RTC_COUNTER, timestamp - d->rtc_bias);
+	return regmap_write(d->regmap, RTC_COUNTER, timestamp - (s32)d->rtc_bias);
 }
 
 static int gamecube_rtc_ioctl(struct device *dev, unsigned int cmd, unsigned long arg)
